requirements/base.txt
Django==5.0.6
django-environ==0.11.2
psycopg2==2.9.9
djangorestframework==3.14.0
django-easy-audit==1.3.6
django-guardian==2.4.0

cryptography==42.0.8

whitenoise==6.5.0
pillow==10.4.0

scikit-image==0.24.0
opencv-python-headless==4.10.0.84
numpy==2.0.0


django-filter==23.5
django-extensions==3.2.3
django-cors-headers==4.3.1
django-storages==1.14.2
drf-extra-fields==3.7.0

drf-spectacular==0.27.2

channels==4.1.0
daphne==4.1.2
gunicorn==22.0.0
setuptools==70.3.0
